# lifecycle-test/src/main/resources/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "lifecycle-test API",
        "description": "OPEN-O Interface provider.",
        "version": "1.0.0"
    },
    "host": "openo.org",
    "schemes": [
        "https",
        "http"
    ],
    "basePath": "/openoapi/vnfsdk/v1",
    "paths": {
        "/vnfpackage": {
            "post": {
                "summary": "Vnfpackage Onboarding",
                "description": "Vnfpackage Onboarding",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Vnfpackage",
                        "description": "Vnfpackage Onboarding.",
                        "schema": {
                            "$ref": "#/definitions/VnfpackageParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Vnfpackage Onboarding",
                        "schema": {
                            "$ref": "#/definitions/VnfpackageOnboardingResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Vnfpackage Onboarding Not Found"
                    },
                    "500": {
                        "description": "Vnfpackage Onboarding failed to process the request"
                    }
                }
            }
        },
        "/vnfpackage/{csarId}": {
            "get": {
                "summary": "Vnfpackage Query",
                "description": "Vnfpackage Query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "csarId",
                        "in": "path",
                        "description": "csarId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Vnfpackage Query",
                        "schema": {
                            "$ref": "#/definitions/VnfpackageQueryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Vnfpackage Onboarding Not Found"
                    },
                    "500": {
                        "description": "Vnfpackage Onboarding failed to process the request"
                    }
                }
            }
        }
    },
    "definitions": {
        "VnfpackageParam": {
            "type": "object",
            "description": "Vnfpackage Param.",
            "required": [
                "csarId",
                "vimIds",
                "labVimId"
            ],
            "properties": {
                "csarId": {
                    "type": "string",
                    "description": "csarId"
                },
                "vimIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "labVimId": {
                    "type": "string",
                    "description": "labVimId"
                }
            }
        },
        "VnfpackageOnboardingResponse": {
            "type": "object",
            "description": "Vnfpackage Onboarding Response.",
            "required": [
                "jobId"
            ],
            "properties": {
                "jobId": {
                    "type": "string",
                    "description": "jobId"
                }
            }
        },
        "VnfpackageQueryResponse": {
            "type": "object",
            "description": "Vnfpackage Onboarding Response.",
            "required": [
                "csarId"
            ],
            "properties": {
                "csarId": {
                    "type": "string",
                    "description": "csarId"
                }
            }
        }
    }
}